More Xcode work. Fix too-small integer warnings.

+++ Source/DOM classes/Core DOM/NamedNodeMap.m
@@ -86,7 +86,7 @@
 
 -(unsigned long)length
 {
-	int count = [self.internalDictionary count];
+	unsigned long count = [self.internalDictionary count];
 	
 	for( NSDictionary* namespaceDict in self.internalDictionaryOfNamespaces )
 	{
